Fix for terminating tempfile error

+++ sqlitedict.py
@@ -296,7 +296,8 @@ class SqliteDict(DictClass):
 
         logger.info("deleting %s" % self.filename)
         try:
-            os.remove(self.filename)
+            if os.path.isfile(self.filename):
+                os.remove(self.filename)
         except (OSError, IOError):
             logger.exception("failed to delete %s" % (self.filename))
